log-collectors/requirements.txt
kafka-python-ng==2.2.3
watchdog==4.0.0
pywin32==306; sys_platform == 'win32'
docker==7.0.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
